UPSTREAM: AGESA: Remove redundant and invalid IRQ routing
The size of the array did not match that of the actual
allocation. Furthermore, the tables are written as
part of set_pci_irqs() in hudson/pci.c.

Also the removed code was never reached runtime, as it is
only executed on ACPI S3 resume path that is currently
disabled.

BUG=none
BRANCH=none
TEST=none

Change-Id: I39342e03eec9c71b2584bb078a6811193e302869
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: 627d7906517db4bff31151584ae397ebc0e39c0a
Original-Change-Id: If1c47d53a7656bdff40d93fc132c8c057184ae46
Original-Signed-off-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/18587
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: https://chromium-review.googlesource.com/451270