UPSTREAM: util/intelmetool: Fix segfault on edge cases
The intel ME checker tool would segfault if it reaches the end of
the loop without having the dev pointer set. This happens when
it gets to the end of the previous loop without knowing what to do
with any of the devices it sees.

This patch makes sure the pointer is not NULL before accessing it.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ibaef27c8be18b6704d60250098a6f676fe197965
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: 2a1ae05316db8d70c8e15df4913b786b120ab615
Original-Change-Id: Ia13191799d7e00185947f9df5188cb2666c43e2a
Original-Signed-off-by: Huan Truong <htruong@tnhh.net>
Original-Reviewed-on: https://review.coreboot.org/18573
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Philipp Deppenwiese <zaolin.daisuki@gmail.com>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: https://chromium-review.googlesource.com/451435