UPSTREAM: chromeos/elog: Filter developer mode entry on S3 resume
The event log entry indicating developer mode is useful for the
boot path, but is not really useful on the resume path and removing
it makes the event log easier to read when developer mode is enabled.

To make this work I have to use #ifdef around the ACPI code since
this is shared with ARM which does not have acpi.h.

BUG=b:36042662
BRANCH=none
TEST=perform suspend/resume on Eve and check that the event log
does not have an entry for Chrome OS Developer Mode.

Change-Id: Ief6dead73856689f0fb0bce6266d66c7196340ee
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: f8401cddb82adbc8c5a8268f6215cf3cf903cfbf
Original-Change-Id: I1a9d775d18e794b41c3d701e5211c238a888501a
Original-Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/18665
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins)
Reviewed-on: https://chromium-review.googlesource.com/452361