tpm: spi: Fix compile-time bug with pointer arguments
CL:452283 (tpm: spi: cr50: try to wake cr50 if it is asleep) introduced
a compile-time error by passing an argument the wrong way (probably due
to cherry-picking from an older branch and not testing it again). Due to
a misconfiguration this slipped by our continuous integration testing.
Fix it.

Jeffy, please re-test with this patch on ToT when you have time so we
can make sure the code actually works as intended now, too.

BRANCH=None?
BUG=None
TEST=Compiled GRU_HAS_TPM2 board.

Change-Id: I3210cd53014a206f5d36abcfe607a9710a8253fa
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/454921
Reviewed-by: Douglas Anderson <dianders@chromium.org>